improved neuron persistence async handling

+++ modules/ViewerController/src/main/java/org/janelia/workstation/controller/action/AnnotationSetRadiusAction.java
@@ -85,7 +85,7 @@ public class AnnotationSetRadiusAction extends AbstractAction {
             return;
         }
 
-        final Float finalRadius = radius;
+        final double finalRadius = radius;
         SimpleWorker setter = new SimpleWorker() {
             @Override
             protected void doStuff() throws Exception {
